move ExtraCheckError to exceptions.py
