+++ backend/ops_api/ops/auth/exceptions.py
@@ -4,3 +4,13 @@ class NotActiveUserError(Exception):
     """
 
     pass
+
+
+class ExtraCheckError(Exception):
+    """Exception used to handle errors from the extra check function that can be passed
+    into @is_authorized().
+    """
+
+    def __init__(self, response_data):
+        super().__init__()
+        self.response_data = response_data
